Add device and platform API routes
